Kfold Example
In k-fold CV, the k-value dictates the number of groups, or “folds”, into which the data will be divided. For example, if k=5 for a dataset with 25 samples, the data will be divided into five groups to form five separate models. In the first model, one-fifth of the data (e.g., samples 1-5) will be designated the test data subset, while the other four-fifths (e.g., samples 6-25) becomes the training data subset. In the second model, a different group of samples becomes the test data subset (e.g., samples 6-10) while the remaining four-fifths of the data (e.g., samples 1-5 and 11-25) are the training data subset. This process is repeated three more times in this example so that each fold containing a distinct set of records is used as the test data subset. When the k-value changes, the number of total folds and their relative size changes.